lun/gateway: update alua support
This patch modifies the current alua support, so we are not
using a single alua group for multiple states. This will
lead to confusion on some OSes.

The patch also drops the preferred bit support because it was not
needed and with how we are currently using a group per iSCSI TPG
leads to confusion on windows.

Signed-off-by: Mike Christie <mchristi@redhat.com>