config, alua, gw: set nonop_delay_msecs to zero
nonop_delay_msecs controls how long the target waits before
executing a IO sent to non optimized path. During the initial
discovery of those paths or failover to those paths, this is
will cause IO to slow down considerably which is not needed
for our model.

This has us set the delay to zero.

Signed-off-by: Mike Christie <mchristi@redhat.com>