config gw: fix alua lun binding
For the SCSI REPORT_TARGET_PORT_GROUPs command we must return the local
ports remote ports. If we have 2 nodes with 1 tpg each, tpg1 should report
AO and tpg2 will be ANO.

In our example, this check

"if this_host == owning_gw"

will be true for both tpg1 and tpg2 and so we are setting both
to active. An RTPG sent to that path will return 2 ports that are AO.
On the other node, it will be false and you get 2 ports that are ANO.

Windows and linux just check the first port, so it works. But,
OSs like ESX will loop over all ports and so they will see
an inconsistent port setup.

This fixes it by checking if the owning_gw's portal_ip_address
is the same as the TPG's ip address. If it is then, we set
the lun under that TPG to AO. If not, then it will be ANO.

Signed-off-by: Mike Christie <mchristi@redhat.com>