Fix for BZ 1384858 - admin updated hosts file but did not provide an ip to group vars
This ensures that when the admin updates the /etc/ansible/hosts file to expand
the gateway group, but *doesn't* provide an additional ip to the gateway_ip_list
variable the condition is caught allowing the request to be failed gracefully. This fix
sets the error flag against the gateway object - a subsequent fix was done to the
ansible module to ensure the error state is picked up.