config.gateway: ensure the ip_list is updated when a gateway is added serially (CLI)
the gateways section in the config object maintains a list of ip's for the portals presented
by each gateway node. With an ansible run, the ip_list is already populated with all IPs for
all gateways, but when using the CLI, IP's are added one at a time. This fix ensures that
each time a new gateway is added, the ip_list gets updated