rbd-target-gw: added sysinfo api, and a configuration watcher
api/sysinfo provides a way the CLI can get system info from other
potential gateways. Authentication is https/username and password

Configuration watcher has been added to track the epoch stamp on the
configuration object - if it changes, the config is refreshed. This reduces
the likelihood of stale data being passed back to the API user