rbd-target-gw: update APIClient class to simplify client creation
The CLI will create a client first and then add auth and disks - but the
api call should allow one call that handles the complete defintion of the
client (iqn, chapp, luns). This update allows the APIClient to use additional
variables passed in to setup the client in one pass.