config.lun: update the gateway's active_lun count when an rbd is removed
Each gateway has metadata that includes a count of the luns that the gateway
holds the primary path for (active_luns). This field is incremented when new
LUNs are defined, and this change decrements the count when a LUN is removed.

active_luns is important to the LUN balancing across the gateways - if it is not
accurate, you can end up with skewed workload across the gateways.