gwcli.client: Add lookup disk->client enabling shared disks to be identified
lun_map added (rbd_path -> set(client_name,...)
This map allows a warning to be triggered when the same rbd image is
mapped to multiple clients. Sharing disks in this manner may be OK - so
this is just flagged as a warning for the Admin to understand that they have
shared the disk across clients.