enable multiple ceph clusters in the UI
The ceph config dir is now read for .conf and keyring files enabling
the creation of multiple cluster trees in the UI. This involves a new
tree called clusters, with each discovered cluster defined as children.

The conf file supplied in the settings file (iscsi-gateway.cfg) is determined
to be the local ceph cluster. This flag then ensures that rados/rbd changes
are only directed to the local cluster. Support for multi-clusters is in prepartion
for rbd-mirror workflows within the cli.