gateway: port open check updated with a shutdown call
The detection method opens a connection to the TCP port iscsi is
listening on. Although the connection is not used, the socket connection
will be retained in a TIME_WAIT state and reaped by the OS later. This change
ois just to send a clear shutdown signal, prior to the socket close call.