client/storage: allow disks to be autodefined when managing a client
rbd's can be defined through /disks create but to keep the amount of switching
around in the UI down, the ability to create disks has been added to the client
module. Now when managing a client, a mapping request for a non-existent disk
will auto-define the disk, and then do the appropriate mapping to the client.