utils.py: added validation functions for disk and gateway requests
validating a disk or gateway request was done within the cli, but this would
mean a straight API call would not be validated leading to unpredictable API
responses. Validation is now done through these functions called by the API
simplifying the cli code and bringing consistency to the CLI and API...at least
for the gateway and disk interactions!