client: lower nop timers
Users want to be able to failover within 15 - 25 seconds, so
we need to detect the dropped conn within that time frame. This
patch lowers the nop timers to 5 and 5, so in a worst case
where the initiator drops the conn right after sending a nop
and that nop takes just under nop timeout seconds to reach us,
and we have to wait another nop interval + nop timeout seconds
to detect the drop, then it will be a 15 second wait. If a command
reaches us in that last second and then gets stuck waiting for
ring memory then there could be another 5 second wait before it
is failed. There is then a less than second delay to kill
everything. So it will take us about 21 secs to cleanup.

Setup notes:

- For linux the iscsi initiator timer
node.session.timeo.replacement_timeout must be set to 25 sec or
longer, and the multipath.conf timer must be set to fast_io_fail_tmo
25 sec or longer. If your version of multipath-tools sets
the replacement/recovery_timeout for the iscsi layer then just
setting the fast_io_fail_tmo is enough.

- For ESX, the "iSCSI recovery timeout" must be set to 25 seconds
or higher.

- For windows, the "LinkDownTime" must be set to 25 seconds or
higher.