node: prevent 'info' command causing an exception
If the object the user is 'at' in the UI does not have a display_attributes
attribute, the get_info method would throw an exception. This fix just
checks for the presence of the attribute and returns with an error
message if the object does not support the info command.