group: host group management fixes
code refactored. The old implementation stored the images for a group as
a simple list, but this does not guarantee consistency across the nodes.

this patch stores the lun-id with the disk within the configuration object.
The group therefore defines lun-id sequence, and the changes to the GWClient
class now allow the definition of a client based on this new image_list
format

More debug text has also been added to track add/remove operations to
groups