group: fixes to process flow
group processing (apply) is split into two steps;
metadata: to handle changes to the config object
client: apply the current disk list to clients within the group

metadata is typically updated by the receiver of the request, which
syncs the group object. Subsequent _hostgroup calls see there is no
difference between the request and the current config state, and bypass
the metadata step, and just apply the lun masking to the gateway's local
LIO instance.