gateway: gateway creation and health checks updated
Creating a gateway now provides a skipchecks option to bypass validity
checking (rpm versions, distro, kernel) when adding a gateway node to
the configuration

Health checks were using a direct socket open against the iscsi target
port - however, this results in errors in the syslog. So health checks
now use the _ping endpoint of the API

In addition to these changes help text has been added to several commands
and a 'refresh' command has been to the gateway entry to check health.