rbd-target-api: health check logic and gateway validity checks updated
GW health is now determined by the _ping endpoint. This endpoint checks
that the iscsi port is open/listening - as long as the gateway already
exists in the config (if it's not listed, 3260 shouldn't be up!)

Validity checks updated with latest version information. However, the
gateway api endpoint also accepts a skipchecks parameter to bypass these
checks. This is intended for development only.