rbd-target-gw: add exception handler to cleanly shutdown on errors
Clearing the config is now in a separate function, and called by an
exception_handler and the halt function. The idea being that whenever an
unexpected error occurs, the config is wipe to return LIO to a known
state.

In addition the definition of the gateway, luns and clients is now done
is separate functions called by the main apply method.