storage: fix the display of capacity when adding disks
The commit data is calculated from the disks within the UI tree. The issue
was that the calc was done *before* the disks were added so the UI
capacity numbers were always one allocation behind. This patch corrects
the sequence to ensure the commit calc is done after the disks are added
to the UI