client: fix issue at boot where luns would not be mapped to clients
Clients defined with a group name, need to ignore LUN requests since being
a member of group means luns are allocated through the group not direct to
the client. However, the catch all was too broad and stopped client
definitions during a gateway restart adding the required luns. This fix
addresses that problem.