fix gateway_keyring setting
The default gateway_keyring is set as the filename and path,
but when the user sets it it can only be the filename because
we merge it with "/etc/ceph".

This fixes the initialization so it is only the filename
and it also has us use the ceph_config_dir instead of
hard coding /etc/ceph.