Cleanup stale locker during restart
If a node dies with the lock held and it restarts before
a failover has executed, then it will look like we are still
holding the lock due to runner not having access to the nonce
part of the client info. This has the service cleanup the stale
locker when it restarts.