Fix node addition
Node addition does not sync existing luns/clients because
it interprets nosync as a bool but request.form.get is
returning a string. The seed_disks/luns functions are then
never called (seed_disk seems to have not been finished as
there is no node=sync for the disk put command).

This has the _gateway put command just call define_clients/disks
added in the ceph-iscsi-config PR to setup the luns and disks
if requested.

v2:
Swtch to exception based.

Signed-off-by: Mike Christie <mchristi@redhat.com>