Support integers as size args for LUN/RBDDev
To avoid converting between the RBD class's bytes size
and human readable format and then back to bytes, have
the LUN and RBDDev class support a integer size in
bytes.

This will fix the bug where during conversion we will
lose precision and hit mismatch errors during startup.

This was debugged by and based on patches by Zhuoyu Zhang
in:

https://github.com/ceph/ceph-iscsi-config/pull/50/
Signed-off-by: Mike Christie <mchristi@redhat.com>