Check for gw manage map errors
The gw manage(map) call calls map_luns which could already fail due
to a LUN() creation error. It can now also fail due to invalid failover
type be set with this patch

commit 6738e54aab2db2345f3660c6403b2b1dbee81646
Author: Mike Christie <mchristi@redhat.com>
Date:   Mon Jul 23 16:45:30 2018 -0500

    Detect and handle invalid alua failover types

This patch has rbd-target-api check for gateway.error being set by a
caller which was not checking for errors. As far as I can tell the other
map callers are already checking for errors.

Signed-off-by: Mike Christie <mchristi@redhat.com>