Remove per client auth nochap support
The requirement that all ACLs have to use the same auth type is
already confusing. The per client nochap command is adding to
that because the user will commonly just disable the one client
or miss a client. This patch removes the per client nochap support
so users have to use the host wide command.

This still does not fix the case where initially nochap is used
and then the user only sets up chap for a subset of clients.

Signed-off-by: Mike Christie <mchristi@redhat.com>