Add host level auth nochap command
This adds a auth nochap command to the host node that resets the auth
settings so chap is disabled and initiator name based ACLs are used.

The problem is that the kernel does not support mixing ACL and chap ACLs
under the same tpg. Users might initially set it up with CHAP then want
to reset back to non chap and they can do that with this command.

In V2 of this patch, gwcli prints out a auth status:

o- hosts ................ [Hosts: 2: Auth: CHAP]

Where Auth can be CHAP, None or MISCONFIG and the latter 2 will be
highlighted in red.

Signed-off-by: Mike Christie <mchristi@redhat.com>