Fix client auth info output
The client auth info output prints UNDEFINED when we haven't setup a
CHAP user/password yet although CHAP is disabled by default. If CHAP is
disabled with the nochap command or via ansible you just see a blank
space in the chap info output.

This patch has us return "None" like is done for the summary when CHAP
is disabled.

Signed-off-by: Mike Christie <mchristi@redhat.com>