Only allow num and alpha chars for image name
Chars like "@" and "/" are not allowable chars for image names. I think
the current regex expression was trying to limit the allowable chars to
numbers, chars, dashes and make sure that there is a "." to separate the
pool and image.

However, we are currently allowing any char like "@#$/" etc.

We seem to need a "^" and "$" to make sure everything between the
beginnging and end of the string is the type listed in the expression
and fail if not. Without them we seem to only check if the image name
has a char, number or dash and as long as it does we return success.

Note: I have no idea what I am doing with regex. I had to google this so
hopefully someone can confirm this works for the reason I think it does.

Signed-off-by: Mike Christie <mchristi@redhat.com>