Move image name check to valid_disk.
The next patches are going consolidate the disk related to checks and
move them to ceph-iscsi-config. This moves the name check to valid_disk.
This function will then be moved to ceph-iscsi-config in the next patch.

This also drops the pool part of the check because we check for a valid
pool a little later in the same function.

Signed-off-by: Mike Christie <mchristi@redhat.com>