mv control code to new class
This moves the control code to a new class so GWTarget, GWClient and
GWLUN ... ehhh ... LUN can all share it.

There are 2 interface changes.

1. The update_*_control functions return errors differently. The acl one
was passing the rtslib exception upwards but the target one was returning
a target.err. This patch syncs on the passing the rtslib exception upwards.

2. control reconfiguration was committed in the target manage function
but for the lun class we had a commit callout. This adds a generic
commit callback so it works the same for all classes.

Signed-off-by: Mike Christie <mchristi@redhat.com>