Fix api default port info
The default api_port in settings.py is 5000 and not 5001.

Note: I think there was a funky ending char in the file. When I edit
this file with a editor in linux it keeps fixing it so there is a extra
change in the patch at the end of the file.