Use ceph-iscsi-config lun check
Drop out check and rely on the valid_disk check added in PR:
https://github.com/ceph/ceph-iscsi-config/pull/98

This fixes a bug where we try to add disk 256 and on the initial node
it works. But because the initial node does a lun.allocate that adds
the disk to the config, and the other nodes will hit the max lun check
which checks the config and fail to add the disk. It then leaves the
rbd image in the config.