packaging: Move python unbuffered option to systemd unit file
This change is required to avoid a problem originated when running
setuptools with the --executable option to override the shebang
of the pyhton script when the original file has a shebang with a
parameter.
Example:
  original file has:

Running python setup.py --executable=/usr/bin/pyhton -s will generate
an executable file with:

which is not a valid shebang.

Signed-off-by: Ricardo Dias <rdias@suse.com>