Fix gw addition to an existing cluster
When we added multiple target support, the gw info was moved from a
global definition to a per target. The problem is that if we are adding
a gw to a cluster with clients/luns then we are doing:

1. Update the existing portals in the target_config.
2. Map luns to the new tpg/portal on the existing gws.
3. Add the new gw to the list of portals in the target_config.

At step 2 we need the new gw's info in bind_alua_group_to_lun so we can
figure out the ALUA groups priority, and we will currently fail since we
have not yet set it up.

This patch fixes the problem by calling _gateway on the new gw first and
then adding all the portalal info at this time. This will also prevent
the target_config from being in a inconsistent state if we fail in the
middle of updating all the gws. Note the next patches will add support
for gw deletion, and so they will add support to fully clean up from
failures when we have only partially updated the cluster during gw
additions.

This also removes the "exists" test in _gateway() since we do
that in valid_gateway from gateway() already. Having it in _gateway
prevents us from adding it early so it can be used later since we will
hit that test when we update the tpg/mappings on the existing gws.

Signed-off-by: Mike Christie <mchristi@redhat.com>