Have target deletion remove disks
We delay storage_object/disk creation until there is a target but when
we delete a target we will leave them in the kernel until the daemon is
stopped. This has us remove the storage objects from the kernel during
target deletion if they are not being used as LUNs in another target
running in the kernel.

Signed-off-by: Mike Christie <mchristi@redhat.com>