Move client/lun define calls to target define function
The gw addition code in rbd-target-api is the same as what we are doing
in gateway.py in the define path except that in rbd-target-api we enable
the network portal while we are still setting up which is wrong.

This patch moves the client/lun define calls in gateway.py to the target
define call so rbd-target-api can call it.

Signed-off-by: Mike Christie <mchristi@redhat.com>