Have define_luns operate on only the target passed in
define_luns was originally created when there was a single target.
When we added multi target support we left it looping over all disks
that are mapped to any targets, and then at the end of the function it
only maps luns for the specific target passed in. So it ends up being a
mix of single and multi target support.

This creates the following problems:

1. If a disk were mapped to multiple targets we could break locks in
use.
2. We call define_luns multiple times during initialization due to
calling it for each target so we can do a map_luns.
3. For new gw creation we end up trying to map luns twice. Once from the
define_luns call then again from the manage('map') call. For the latter
we only needed the manage call for the update_acl call. And then it
seems that during daemon initialization we did not do a manage call
so update_acl might not have been called.

This patch fixes these issues by moving the update_acl call to when we
call manage('target'). Users then only need to call define_luns. That
function is then fixed to only operate on the target passed in and only
break locks if the disk is not mapped to other LIO luns at the time.

Note: manage('map') is not used and will be removed in future patchsets
that removes all ansible support.

Signed-off-by: Mike Christie <mchristi@redhat.com>