rbd-target-api/gw: Fix systemd deps
rbd-target-api and rbd-target-gw used to both be able to modify the
system, so we bound api to gw so if gw went down api would be stopped
too. This prevented races where if gw started back up api could also be
modifying the system at the same time.

When we moved the target start up code to api, we no longer had this
issue, but I forgot to update the systemd deps. This patch modifies them
so:

1. gw is bound to api. This is not required to prevent modification
races like before, but if api is stopped then the system will not have
any running target so I do not think there is no point in having gw up.

2. gw is now sarted after api. We used to need the reverse because gw
did the initial setup. Now gw needs the system setup to handle statistic
calls.

3. api is started after tcmu-runner

Signed-off-by: Mike Christie <mchristi@redhat.com>