Update spec to use systemd scriplets
This updates our spec to use systemd scriptlets for fedora/RHEL.
They are described here:

https://docs.fedoraproject.org/en-US/packaging-guidelines/Scriptlets/

There is one major behavior change with this patch. I followed the doc
assuming we wanted to get into fedora/RHEL. The problem is that this
doc:

https://docs.fedoraproject.org/en-US/packaging-guidelines/DefaultServices/

lists requirements for enabling the service when the package is
installed and we do not meet them, so this patch does not enable us
by default.

However, I can just update the docs for this. The upstream/RHCS docs
already state we must enable rbd-target-api so we just need to add
a line about also enabling rbd-target-gw.

Signed-off-by: Mike Christie <mchristi@redhat.com>