rbd-target-gw/api systemd dep fix
Newer versions of systemd will return error when trying to start
rbd-target-gw, because it detects ordering and dependency issue
with rbd-target-api.

This removes the bind/wants between the 2 services. The bind/wants was
not required and was added to make it more convenient for users so that
the services stopped and restarted together.