Fix openssl rpm dep
rbd-target-api requires ssl, so add it to the spec file.

There is some weirdness here in that in RHEL 7 the package is named
pyOpenSSL, but in suse and RHEL8 and modern Fedora versions the package
was renamed to python-pyOpenSSL/python2-pyOpenSSL for the python2 based
package. In python3 it is always name python3-pyOpenSSL for all distros.