Fix image creation help
Fix create/attach help. We no longer support max_data_area_mb during
creation because we do not create the kernel device until the image is
mapped to a LUN, so drop the max_data_area_mb reference.

The create command only creates a RBD image. It does not yet map the
image to a LUN, so replace LUN with image.

Signed-off-by: Mike Christie <mchristi@redhat.com>