Fix image deletion.
If we have created a image, but not yet mapped it then remote gateways
may not have updated their config, or if the user switches gws the
command is run on then the now local gw may not have updated it self.
Add a refresh to the disk and _disk commands to make sure it is up to
date.

Signed-off-by: Mike Christie <mchristi@redhat.com>