Fix config upgrade from 9
This fixes 2 bugs:

1. While we are in version 9.5, the config watch would detect the
version being updated to 9.5 then swing back on us. We only check for
full version upgrades so we would end up updating the version again and
looping like this until the other nodes were updated.

2. If upgrading from 9 to version > 10, the new versions added fields
that were then later refefenced when setting up the target. If all the
gws did not upgrade quickly enough, like with a rolling upgrade, then
the config watch might not have detected the config change and updated
us to 11 and added the new fields. We would then crash.

This patch skips the intermediate version 9.5 and checks if the node is
in the gateways_upgraded list or not to determine if the node needs to
update it's gw name fields still.

This is very fragile since it relies on no future upgrades needing the
updated fqdn based config values. However, I am not sure if it's worth
it to add code to sync our state at this point if we do not need it. If
we do end up needing this, then it might be better to have users upgrade
to a version 10, then from there upgrade to the current version.

Signed-off-by: Mike Christie <mchristi@redhat.com>