Fix crash when CephCluster fails to get Rados object
If the Rados() fails the __init__ function returns an error. The caller
then prints out the error_message and __del__ is called. This then
results in a crash because self.cluster failed to get set. This fixes
the bug by initializing it to None. It also adds a more descriptive
error message.