gwcli/api: add force option for gw removal
We currently remove the gateway from the config on the gateway we want
to remove. If the gw is down permanently this creates the problem that
we cannot delete the gw unless you re-create it. This patch adds a force
option, so gwcli and api users can tell rbd-target-api on the local node
to just update the config there and ignore the dead gw.

Signed-off-by: Mike Christie <mchristi@redhat.com>