Fix metrics crash if gw not setup yet
If no targets are setup or the target is not mapped to a gateway,
then the rbd-target-gw metrics call will crash due to trying to
access target.wwn or a empty list of tpgs. This just adds some
checks to make sure we have a target and that its mapped to a
gateway.

Signed-off-by: Mike Christie <mchristi@redhat.com>