Fix crash during unblacklisting with python3
In python3 check_output returns result as a byte type, where in
python2 we got a string. In python3 there is an arg for check_output
to control the encoding, but we do not have it in python2. So this
has us rely on the the "ceph osd blacklist ..." command returning a
proper error code if it fails instead of parsing the output.

Signed-off-by: Mike Christie <mchristi@redhat.com>