Always perform full deletion on local node
If we are running the gateway deletion command always try to do the full
(lio and config) deletion if the gateway to be deleted is the local
host. The force option where we only clean up the config config only
makes sense if a remote node is dead. If we only cleanup the config
object on the local node, then we would leave the target running in the
kernel when we can always clean it up.

Signed-off-by: Mike Christie <mchristi@redhat.com>