Fix ceph blacklist ls test
I am not sure if it could always happen or it is a change in python3,
but while testing the blacklist clean up code with python3 we
sometimes do not see the blacklist get fully cleaned up. The problem
is that sometimes blacklist.decode() will change the order of the
returned items so the "list N entries" string. When we look over the
output we then skip the first one and it does not get cleaned up.