Fix resolve_ip_addresses formatting
This has resolve_ip_addresses return addresses in the format expected by
the caller. Specifically, requires_restricted_auth expects that the
interface scope and leading "::ffff:" are dropped.

Note:
- I am not 100% sure when we even see this. We have had a couple users
  hit the issue though. One known case is when the user mis-sets the IP
in the /etc/hosts file. In this case if you just added an extra dot:

192.168.1..100 somehost.local

then getaddrinfo will return both the link local ipv6 addr and a
corrected ipv4 addr. If the address is correct though, then we get the
ipv4 address only.

In other cases, we have not seen any misconfigs, so I am guessing it
might be differences in python versions, or expected behavior in certain
configs. However, I am not sure what those are right now.