Add a config option to log to stderr
Most the other ceph components have an option to get them to log to
stderr. And this is used when containerising to help gather logs.

This patch adds a config options `log_to_stderr`, `log_to_stderr_prefix`
and `log_to_file` to do the same for ceph-iscsi. These are matching the
other ceph daemon's:

  --default-log-to-stderr=true --default-log-stderr-prefix="debug "
  --default-log-to-file=false

Signed-off-by: Matthew Oliver <moliver@suse.com>