ceph-iscsi: add erasure pool support
The erasure coded pool does not support omap, here we will just store
the data in it and will store the metadata in a replicated pool.

This will add a "datapool=<ec_pool_name>" parameter to specify the
erasure coded pool to store the data when creating a disk, and the
"pool=<name>", which is "rbd" as default, will be used to store the
metadata only.

The backstore object name is <pool>.<image> in the replicated pool,
which is the same with the none-erasure image.

Signed-off-by: Xiubo Li <xiubli@redhat.com>