lun: try to recovery the config from the LIO
If the LIO exists but the config doesn't, it will only add a disk
item with "created":

"rbd/blockX1": {
    "created": "2022/02/22 11:23:12",
}

We can recovery the config from LIO instead of just append a corrupted
disk item.

Currently we won't support the size changing case and if the image
size and LIO sizes don't match we will skip recovery.

Signed-off-by: Xiubo Li <xiubli@redhat.com>