spec: make sure we always have the target dir is
rtslib now requires /etc/target or /var/target to be setup (it
depends on the version). The issue is that rtslib does not make
those dirs and instead relies on targetcli or ceph-iscsi to make
them.

For container cases we do not install targetcli, so ceph-iscsi's
daemons will crash or fail to setup disks, when we make rtslib
calls that try to access the etc/var target dirs.

When rtslib is fixed drop this and update the dep. Note that we
add both etc and var, because the kernel and userspace developers
keep switching the dir they want to use.

Signed-off-by: Xiubo Li <xiubli@redhat.com>